Add background process, update CLI
the goal of the background process is to upload any changes to a
directory.

Usage: `owo-bg -k MY_KEY -p PATH`
Default path is the `screenshot` directory on android phones.

The main purpose here was to make android uploading a lot easier.

On mobile, this HAS TO BE RUN on [`termux`](https://termux.com/)
This commit will send push notifications and copy to clipboard if
`termux-api` is installed.